// src/shared/ui/stack/types/stack.ts
export type FlexJustify =
    | 'justifyStart'
    | 'justifyEnd'
    | 'justifyCenter'
    | 'justifyBetween';
export type FlexAlign = 'alignStart' | 'alignEnd' | 'alignCenter';
export type FlexDirection = 'row' | 'column';
export type FlexGap = 'gap2' | 'gap4' | 'gap8' | 'gap16' | 'gap32' | 'gap100';